✨ feat(extensions): add days and daysAgo to Int
Extend the Int extension to include `days` and `daysAgo` properties. This allows for more convenient creation of `Duration` objects representing days and for calculating dates in the past based on a number of days.

This change also refactors the preview data generation in `EpisodesListView.swift` and `SelectablePodcastsGridView.swift` to utilize the new `daysAgo` functionality for more realistic sample data with varied publication dates. The episode creation loops in these previews have been adjusted to generate more episodes with these date variations.