✨ feat(PodcastsGridView): Add subscribed badge to podcast grid items
Display a visual badge on podcast grid items to indicate to the user if they are subscribed to that podcast. This improves the user's ability to quickly identify subscribed content within their podcast library.

The `subscribedBadge` private view modifier is introduced to encapsulate the styling and appearance of the badge. This badge is then conditionally overlaid on the `SquareImage` when the `podcastWithEpisodeMetadata.subscribed` property is true.