♻️ refactor(protocols): Simplify fetching and creating podcasts/episodes
Removes redundant `getOrCreatePodcast` and `getOrCreatePodcastEpisode` implementations from base protocol extensions and provides specific default implementations for Concrete types where the object already exists. This streamlines the code by avoiding unnecessary fetches when an object is already available in the current context.

The `PodcastsListViewModel` has also been updated to remove the direct call to `getOrCreatePodcast(_:)` as it's no longer necessary given the changes to the `ManagingPodcasts` protocol. The `selectedPodcasts` computed property in `PodcastsListViewModel` was also moved to `SelectablePodcastList` as it is more appropriate there.